requirements.txt
fastapi==0.99.1
uvicorn==0.23.2
python-multipart==0.0.6
pytesseract==0.3.10
opencv-python==4.8.0.76
numpy
Pillow==10.0.0